missed-packet timeout doesn't make sense
The observed behavior for me is that when sync gets lost, all packets
are lost forever.  There's no point waiting for multiple trips around
the frequency carousel.  Repeated misses on a given channel are most
likely an indication of noise on that channel, so resetting the receiver
is useless in that case too.  Finally, even if counting per channel is
desired, the for loop here is useless; it checks all of the channel-miss
counters after incrementing exactly one of them.