add sanity buffer to discard bad packets when sync is lost
The behavior I see is that when sync is lost, the last packet "decoded"
is actually garbage.  The idea here is to create a queue for incoming
packets, and only log them when they reach queue age N.  The queue is
discarded when sync is lost.  With luck, the garbage packets will be in
the queue and never logged.